+++ src/slice_header.rs
@@ -444,6 +444,8 @@ mod tests {
             num_h_slices: Some(2),
             num_v_slices: Some(2),
             quant_table_set_count: Some(2),
+            ec: Some(0),
+            intra: Some(false),
         }
     }