+++ tests/rgb_encode_frame.rs
@@ -49,6 +49,8 @@ fn rgb_v3_cr(
         num_h_slices: Some(num_h),
         num_v_slices: Some(num_v),
         quant_table_set_count: Some(1),
+        ec: Some(0),
+        intra: Some(false),
     }
 }